Implement list initializers.
Implement list initialization of vectors, matrices, arrays and classes.
This required introducing a new non-terminal in the parser, expr-or-list, since
list initializers are not allowed in ExprStmt (this would be a parsing ambiguity).
The type system also has a representation for an unresolved list initializer (ListType),
so that we can match it during method matching.  ListType implements CanWidenTo() by
querying the calling type if it CanInitFrom() the list type. Each type that can be
widened from a list initializer implements CanInitFrom() as appropriate.
Note that list initializers can also be widened to a weak pointer. This is performed by
creating a stack temporary, and returning a pointer to it.

The semantic analysis for list initializers is done in two passes. On
first encountering an UnresolvedInitializer, the caller may not yet
know what type it expects (e.g., we may have not yet done method matching). So simply
clone it. For all the nonterminal rules that support list initializers, make a call
to Widen(). This will either add a necessary cast, or if it's a list initializer, replace
it with the appropriately-typed expression.

Update many tests and samples to use list-init.