Rework Texture class hierarchy.
This change reworks the texture class hierarchy to reflect the
Compatibility Mode limitations in WebGPU. (It doesn't, however,
enable Compatibility Mode in Toucan.) In particular, a texture
must be allocated up-front with the type of sampling which will be
performed from it.

The five Texture classes are now 1D, 2D, 2DArray, 3D and Cube. They
can only create Sampleable views of the corresponding type.
CubeArray has been removed, since it's not available in Compatibility
Mode.

Support for Constructor and Arg nodes was added to the DumpAsSource
pass, in order to allow default values for the origin parameters
to CopyFromBuffer(), many of which are vectors.

Turn skybox.t into a sample.