Class initialization; zero initialization.
Implement initialization of classes via inline assignment. Add a
defaultValue to fields in order to initialize them. Implement
initialization of classes via field assignment to the default value.

Add a zero init statement, and use it for all variables which don't have
a default value. Implement the zero init statement in LLVM. In SPIR-V,
implement zero-init as a no-op, since all local variables are already
zero-initialized.