Move class constructor handling to semantic pass.
Previously, constructor calls for stack-allocated classes were
handled with a method "sidecar" on the ConstructorNode that
was handled in bespoke fashion by the LLVM codegen (and not at
all by the SPIR-V codegen).

Instead, handle constructor calls in the semantic pass: stack-allocate
the class object as normal, wrap the result with a MethodCall to the
constructor, and return that method call during resolution of the
ConstructorNode. Remove the sidecar from ConstructorNode and the
sidecar handling from LLVM codegen.

The return value for constructor methods was changed from a StrongPtr
to a WeakPtr. This avoids incorrectly attempting to free the result.
(The LLVM codegen didn't have this problem, since it didn't allocate a
temporary to clean it up).

Also fix LLVM codegen for class initializers to use InsertValue,
same as matrix and array.