Implement TempVarExpr and RawToWeakPtr.
Split the AddressOf AST node into TempVarExpr and RawToWeakPtr.
This makes separates responsibilities, and means the codegen backends
don't have to handle both RawPtr and non-RawPtr arguments to AddressOf.
TempVarExpr is responsible for allocating a temporary variable on the
stack, and (optionally) initializing it. RawToWeakPtr is responsible
for creating a control block and returning a weak pointer.