Include the parsed file's directory by default.
In order to simplify includes, the include statement will now
search for the file in the same directory as the source file,
similar to #include "" in the C preprocessor. The default search at
the current working directory was removed.

Also, the cache of previously-included files were being keyed by
filename, causing an error if the same file was included via different
paths. Key it by path instead, and add a test.