Implement Android support.
Enable ARM and Aarch64 support in LLVM codegen.

Add GN support for the Android SDK toolchain.

Use Android's ANativeActivity and android_native_app_glue to
avoid adding any Java requirements to the Toucan build.

Split the TARGET_IS_* macros to TARGET_OS_IS_* and TARGET_CPU_IS_*
macros, initialized from the corresponding gn variables.

Add the appropriate target triples to tc.

Add X86, ARM, and AArch64 explicitly to recommended CMake args.

Capture results of run-llvm-config with ARM and AArch64 enabled.

Remove a spurious dependency on dawn:proc_static.

Implement the basics of the Toucan API on Android: Window, Device,
SwapChain, System.GetCurrentTime(). Events are not yet supported.

Generate ndk_host and ndk_target variables from host_os and target_cpu.

Introduce the toucan_app macro for building apks on Android. These are
simply executables on other platforms for now, but could be apps on
MacOS, for example. Build all samples as Toucan apps.

Factor a toucan_objects GN macro out of toucan_executable that builds
the Toucan .o's.

Link all toucan-generated native code into a single .so for Android,
linking statically against Dawn and libc++. This avoids the need to
package up any additional .so's.

Implement an android_main() to interface with the native glue code.
This is linked into the top-level .so in place of the main() call.

Implement a make-apk.py script to generate an APK from the .so, and
sign it with a checked-in key. It also generates an Android manifest.