Update to Surface-based SwapChains.
Move the wgpu::Surface from Window to SwapChain.
WASM is not yet updated, since it requires an Emscripten update.

Additional fixes:
Use wgpu::PresentMode::Fifo on Android and X11. Update the "skybox"
and "shiny" samples to process all events before a redraw (this is the
only reason we were using Immediate mode on Vulkan).
Fix error logging on all native platforms: pass the DeviceDescriptor to the
CreateDevice() call, and use __android_log_print on Android.