Implement System.GetScreenSize().
Implementation is straightforward on all platforms.
Update the Triangle and Window samples to use GetScreenSize().
Update the Shiny and Skybox samples to use GetScreenSize() and handle aspect
ratio.
X11 fix: store and return the requested size, not the actual size of
the window.
Android fix: Window.GetSize() height was unset.