Fix SPIR-V call graph bug.
If the call graph depth was > 1, and the nested call was not
also called from the top level, SPIR-V code gen would generate it.

Fixed by maintaining a list of pending methods in CodeGenSPIRV,
rather than using the functions_ map.

Remove all the placeholder "priming" in samples.

Add a test.