Implement ExprWithStmt.
This expression node contains an expression and a "sidecar" statement.
It evaluates the expression, executes the statement, then returns the expression's
value. For example, it's useful for doing temporary stores in the middle of an
expression.

It's used here to move the pre- and post- increment/decrement from the
codegen passes to the semantic pass.