SPIR-V: fix pointer to aggregate subcomponents.
SPIR-V prohibits passing pointers to array elements or struct members in
local functions. This is fixed by storing the element to a temporary
(if required), passing a pointer to the temporary, then storing it
back (if required).

A new pass, the ShaderPrepPass, takes on this resposibility when
visiting method calls. It also handles some other shader-isms, such as
the fact that smart pointers are actually all raw pointers in SPIR-V.

Since ShaderPrepPass is a CopyVisitor, and the first one that runs
after the SemanticPass, some additional node copying was implemented in
CopyVisitor.

Finally, a fix was required to the Dawn buffer usages in order to allow
readback from a host-side buffer.

A test exercises the array and structure cases.