Fix increment statement with side effects.
Problem: increment statements re-use the RHS expression, causing it to
be visited twice, code-gen'ed twice, and side effects to be run twice.

Fix: cache all copied nodes in the CopyVisitor, and all Value nodes in
LLVM codegen. (TODO: SPIR-V. This cannot currently be repro'ed.)

Nodes with the copyFileLocation_ flag set to "false" are not cached,
since they must be visited uniquely to use the caller's file location.