Implement raw pointers.
Raw pointers (already used internally by the Toucan compiler) are
now exposed as first class datatypes in the Toucan language in a manner
similar to C++ references.

Raw pointer variables:
  - can never be null
  - must be initialized on creation
  - can never be reassigned
  - cannot be stored in a heap-allocated object
  - cannot be members of an aggregate (for now)
  - are automatically dereferenced (loaded) when used as an r-value

They can be created by taking the address of a stack-allocated value (via
the & operator), or by casting a strong or weak pointer to raw
pointer. The latter will increment the refcount on the containing
object until the end of the scope containing the raw pointer variable,
ensuring it remains valid.

Raw pointers to runtime-sized arrays (&[]) are represented in codegen as fat
pointers, containing the data and length. Values of the following types
may be cast to raw pointers to runtime-sized array:
  - raw pointers to fixed-sized arrays
  - strong or weak pointers to arrays (fixed-size or runtime-sized)
  - vectors
  - matrices
  - list expressions

At the codegen level, array access and length expressions only operate
on raw pointers to runtime-sized arrays. All other types for indexing
are converted to &[] during the semantic pass (in MakeIndexable()).

The SmartToRawPtr node is now also responsible for converting smart arrays
to raw arrays during codegen. The length is extract from the control block
and a fat pointer is created with it.

Since the & operator now produces RawPtrs, not WeakPtrs, raw pointers
are now used where weak pointers were used in RenderPass, ComputePass,
and BindGroup.

Memory auto-layout was moved before the semantic pass. This is
required so that array memory layouts are set correctly during
semantic analysis.

Bindings generation now writes out MemoryLayout for classes and arrays.

LLVM cast creation was cleaned up, now that it no longer has to handle
pointer special cases.