Rework heap allocation and initialization.
Implement brace initialization for the "new" operator.
A call to "new" without parentheses or braces is now
brace-initialization.
If parentheses are used, it is a constructor call, and a corresponding
constructor must exist.

Remove the NewExpr and NewArrayExpr AST nodes in favour of more
fine-grained HeapAllocation and RawToSmartPtr nodes.
"new" operator with a constructor is implemented as:
  RawToSmartPtr(MethodCall(constructor(HeapAllocation, args))).
"new" operator with brace initialization is implemented as:
  RawToSmartPtr(ExprWithStmt(HeapAllocation, stmts)).
where "stmts" are the initialization statements.

Native constructors no longer have special handling at the parser or
semantic pass level. They receive the same AST nodes as above.
CodeGenLLVM::GenerateMethodCall() then detects native constructor calls,
extracts the qualifiers from the HeapAllocation type, and removes the
HeapAllocation parameter. Skipping the first arg must also be done in
bindings generation and method name manging.

Add a flag to UnresolvedNewExpr to indicate whether it's a constructor
or brace initializer.
Rename SemanticPass::InitializeVar() to SemanticPass::Initialize(),
and use it for "new" operator brace initialization.
Change CodeGenLLVM::ConvertFromNative() to only automatically add
control blocks for strong and weak ptrs. Constructor calls (that
return raw ptrs) will use the same RawToSmartPtr node already in the AST.
The parser was refactored to use common initialization non-terminals
for both expressions and the "new" operator. This allows all types
(not just simple types) to be brace-initialized as well as new-allocated.

Update to samples and types, largely to remove of now-incorrect brackets
on "new" operator calls.