Implement array initialization.
Implement array initialization for stack and heap values.
Due to the runtime bounds, this requires a for-loop, generated during
the semantic pass.
A zero-length initializer causes default initialization of each element.
A one-length initializer is "splatted" to all elements.
Any other length of initializer is access via ArrayAccess, and will be
subject to its bounds checking.
These semantics are applied to arguments to the "new" operator, as well
as stack-constructed values and list initializers.
Apply these same semantics to vectors and matrices.
Remove vector splatting from SPIR-V codegen (now redundant).