Fix DepthStencilState binding.
Classes should be written out as ClassName, native or not. This fixes
raw ptrs to non-native classes, such as DepthStencilState.
However, smart pointers to most things should be written out as Object*,
with the exception of smart pointers to native classes. This mimics
the logic in CodeGenLLVM's ConvertToNative() and ConvertFromNative().