Implement list expr default arguments in bindings.
Implement support for list expressions as default arguments in native
bindings.  For now, only empty lists are supported ({}).
Use it to provide a default argument for RenderPipeline's
blendStencilState.
Since templated functions (such as the RenderPipeline constructor) are
not resolved until after type reloading, this required support for
UnresolvedListExpr and ArgList in bindings generation.
Move primitiveTopology to the first argument of the RenderPipeline
constructor, since it's more commonly non-default.
Fix all samples and tests.