Fix list initialization of aggregated classes.
When resolving a list expression to a class, if a field does not
have a default value and is itself a class, create a default
initializer for it. This ensures that aggregated classes are also
at default values.
Refactor this work into an AddDefaultInitializers() function, so it
can be used by the enclosing as well as the aggregated class.
Perform this initialization for empty list expressions in addition
to list expressions with named arguments.
Add tests.