Use a placeholder type for ListType in bindings.
ListTypes for large initializers can be quite large, and
even cause the WASM backend to fail compile with "too many locals".
In the future, only types reachable from generated code (via TypeID)
will be emitted in the bindings, and this won't be a problem, since
user-specified ListTypes should never be reachable.
For now, generate a unique placeholder for every ListType.