Make assignables (r-values) always RawPtr.
Previously, some assignable nodes could return non-RawPtr,
notably method calls and the length expression. As a result, the
"load" inserted by the assignable -> expression reduction rule had to
be a no-op in some cases. There were also many special cases in
field access, array access, and method calls to handle rawptr, non-rawptr,
rawptr-to-smartptr, non-rawptr-to-smartptr, etc.

The solution is to make all assignables (r-values) RawPtrs. If the
result of a method call is non-rawptr, insert a TempVarExpr to spill
it to the stack. Same for the length expression, and enum value access.
If the result is subsequently loaded, the load instruction will "eat"
the TempVarExpr and simply return its value. Field access, array
access, and method calls become much simpler, since they can now
assert that their input is a RawPtr, removing many of the special
cases and still permitting chained access.

Semantic analysis:
Allow RawPtrs to compatible types to be cast. This allows the a qualified
"this" pointer to match an unqualified one.
Refactor auto-dereferencing for field access, array access, casts, and
method calls into an AutoDereference() helper.
AST:
Simplify ArrayAccess::GetType(), since it only needs to support rawptr
to runtime-sized array.
Shader prep:
Remove qualifiers when checking if a type needs unfolding. This was a bug.
Implement copying of TempVarExpr in the CopyVisitor.
Parser:
Make initializer into an assignable. This allows it to be chained.