GenBindings followup.
Split InitTypes() into two flavours: InitTypes() and InitAPI().
The former is used by applications, and loads only types, not
symbols or AST nodes. The latter is used by the compilers, and
initializes the API symbols and types for use in compilation.
Rename emitScopesAndStatements to emitSymbolsAndStatements for
accuracy.
In EmitType(), don't generate the type's ID until write time.
This numbers the types in order of appearance, but has no other
effect on functionality.
Only generate and return the (static) typeList for InitTypes().
InitAPI() does not need to look up the referenced types.