Unify resolution of list exprs and initializers.
Unify the resolution of list expressions ({}) and initializers
(C{}).
Make ResolveListExpr take an ArgList instead of an UnresolvedListExpr.
Make resolution of an UnresolvedInitializer handle only the class
constructor special case, and otherwise call ResolveListExpr.
For class initializers, this ensures we always get an exprlist of
length equal to the total number of class fields (even if some of
them are null). Assert that in LLVM codegen.