Image decoder fixes.
Rename ImageDecoder -> Image.
Make the Image constructor take a strong pointer (and a ref) on the
encoded data. Free it in the Image destructor.
If the input buffer is null, return a null Image.
Make Decode take a raw pointer to runtime-sized array for the decode
buffer, putting the onus of null checking on the caller.
Add support for 1-component JPEGs into RGBA8 results.