Port to iOS.
The minimum deployment target is currently iOS 15.0.
Implement support for .app files and icons on both MacOS and iOS
(tools/make-app.py).
Implement support for .ipa files on iOS (tools/make-ipa.py).
The "toucan_app" template now generates an .app file as top-level target
on MacOS, and an .ipa file as top-level target on iOS.
Only a single Window is currently supported on iOS.
System.Print*() as well as WebGPU error messages are implemented via
os_log() calls.
Implement an "api_ios.mm" that also includes main().
Similar to the Android port, all Toucan calls are run on a separate
thread in order to implement a blocking event loop.
The main thread creates all of the UIKit objects and dispatches events
to the Toucan thread via mutexes and condition variables.