Implement slice operator.
The syntax is array[start..end].
Like Python, the lower bound is inclusive while the upper bound
is exclusive. Both are optional: "start" defaults to zero, and "end"
defaults to the array size.
The operator takes an expression of type raw pointer to runtime-sized array,
but all array and vector types can be implicitly converted to same.
The return value is also a raw pointer to runtime-sized array.
Slices are prohibited in shader methods, and are validated out.