Scope handling rework
Move scoping responisibilities to the Stmts node and modify
SymbolTable to become a simple stack of Stmts.

The compilers no longer create a root scope or symbol table, but a
root Stmts instead, passed into InitAPI() and ParseProgram().
SymbolTable is now instantiated as needed by the parser and the
semantic pass.

Stmts maintains an ID -> Expr map for identifiers and an ID -> Type
map for types. Stmts are pushed and popped off of the symbol table
during both parsing and semantic analysis. Types are added during
parsing and IDs are added during the semantic pass.

IDs may be local variable access (ID -> VarExpr) added when visiting a
VarDeclaration, field access (ID -> this.fieldname) added when
visiting an UnresolvedClassDefinition, or formal argument access
(ID -> VarExpr) added when resolving a Method.  These IDs are looked
up when visiting an UnresolvedIdentifier.

The semantic pass now maintains a currentMethod_, used when visiting a
return statement to validate the return type as well as to know how
far to unwind the stack. This relieves the symbol table of knowing
which type of scope is current (class, method, enum, or stmts).

UnresolvedClassType now holds a ClassType instead of a Scope. During
parsing, a list of statements is built for the class body, consisting
of VarDeclarations, and MethodDecls.  At the end of a class
declaration, these statments are converted into Fields and Methods via
the ClassPopulator visitor. For now, MethodDecls are only used during
parsing and are not encountered during type replacement or semantic
analysis.

Variable declarations produce a Decls node, which is a scopeless list
of VarDeclarations (actually Stmts). Previously, a Stmts node with a
null Scope was used for this purpose.

Enums, formerly using the symbol table for scoping, now use their
own currentEnumType_ global in the parser.

The compilers now create a root Stmts node instead of a Scope, and
pass that to all passes.

Since the symbol table is no longer persistent and doesn't own its
scopes, SymbolTable::Dump() and the associated command-line flag were
removed.